Make some changes to backend
