+++ backend/README.md
@@ -1,0 +1,3 @@
+# backend
+
+Start backend with `npm run dev`
