+++ backend/package.json
@@ -1,19 +1,24 @@
 {
-  "name": "backend",
-  "version": "1.0.0",
-  "description": "",
-  "main": "index.js",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
-  },
-  "keywords": [],
-  "author": "",
-  "license": "ISC",
-  "dependencies": {
-    "@sendgrid/mail": "^6.4.0",
-    "cors": "^2.8.5",
-    "express": "^4.17.0",
-    "live-server": "^1.2.1",
-    "node-cron": "^2.0.3"
-  }
+    "name": "backend",
+    "version": "1.0.0",
+    "description": "",
+    "main": "index.js",
+    "scripts": {
+        "dev": "nodemon index.js"
+    },
+    "keywords": [],
+    "author": "",
+    "license": "ISC",
+    "dependencies": {
+        "@sendgrid/mail": "^6.4.0",
+        "cors": "^2.8.5",
+        "express": "^4.17.0",
+        "node-cron": "^2.0.3"
+    },
+    "prettier": {
+        "tabWidth": 4
+    },
+    "devDependencies": {
+        "nodemon": "^1.19.1"
+    }
 }
